Moved Views into View Folder
